Use kiosk-browser-control interactive to open an interactive Shell within a running kiosk-browser session.
Chromium on Raspbian does not support --start-fullscreen so we try
to use --app instead and instrument openbox to maximize the window.

openbox needs to be killed with -9 (at leat on Raspbian), try to catch
the case where a sleep 313 is left over (maybe because a new sleep 313
starts just when we terminate).